# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/gift/collection/:page": {
            "get": {
                "description": "Search gifts page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Search a collection of gift given a page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page to be retrieved",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Gift"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/gift/create": {
            "post": {
                "description": "save a new gift",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Save gift",
                "parameters": [
                    {
                        "description": "gift object",
                        "name": "gift",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Gift"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/gift/search/:id": {
            "get": {
                "description": "Search gift identified by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Search gift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gift id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gift"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Gift": {
            "type": "object",
            "properties": {
                "boleto": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pixCode": {
                    "type": "string"
                },
                "productImage": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}